+++ packages/isaacscript-cli/file-templates/static-ts/.eslintrc.template.cjs
@@ -1,14 +1,15 @@
 // This is the configuration file for ESLint, the TypeScript linter:
-// https://eslint.org/docs/latest/user-guide/configuring
+// https://eslint.org/docs/latest/use/configure/
 module.exports = {
   extends: [
-    // The linter base is the IsaacScript mod config:
-    // https://github.com/IsaacScript/isaacscript/blob/main/packages/eslint-config-isaacscript/mod.js
-    "eslint-config-isaacscript/mod",
+    // The linter base is the shared IsaacScript config:
+    // https://github.com/IsaacScript/isaacscript/blob/main/packages/eslint-config-isaacscript/base.js
+    "eslint-config-isaacscript/base",
   ],
 
-  // The ".prettierrc.cjs" file is ignored by default, so we have to un-ignore it.
-  ignorePatterns: ["!.prettierrc.cjs"],
+  // The ".prettierrc.cjs" file is ignored by default, so we have to un-ignore it. Additionally, we
+  // don't bother linting the compiled output.
+  ignorePatterns: ["!.prettierrc.cjs", "**/dist/**"],
 
   parserOptions: {
     // ESLint needs to know about the project's TypeScript settings in order for TypeScript-specific