+++ packages/isaacscript-cli/src/commands/init/createProject.ts
@@ -1,5 +1,9 @@
 import chalk from "chalk";
-import { repeat } from "isaacscript-common-ts";
+import {
+  removeLinesBetweenMarkers,
+  removeLinesMatching,
+  repeat,
+} from "isaacscript-common-ts";
 import path from "node:path";
 import { Config } from "../../classes/Config.js";
 import { createConfigFile } from "../../configFile.js";
@@ -14,9 +18,9 @@ import {
   METADATA_XML,
   METADATA_XML_TEMPLATE_PATH,
   PACKAGE_JSON,
-  PACKAGE_JSON_TEMPLATE_PATH,
   README_MD,
-  README_MD_TEMPLATES_PATH,
+  TEMPLATES_DIR,
+  TEMPLATES_DYNAMIC_DIR,
   TEMPLATES_STATIC_DIR,
 } from "../../constants.js";
 import { PackageManager } from "../../enums/PackageManager.js";
@@ -30,16 +34,19 @@ import {
   getPackageManagerNPXCommand,
 } from "../../packageManager.js";
 
+const TEMPLATE_MOD_MARKER = "@template-mod";
+
 export function createProject(
   projectName: string,
   authorName: string | undefined,
   projectPath: string,
   createNewDir: boolean,
-  modsDirectory: string,
-  saveSlot: number,
+  modsDirectory: string | undefined,
+  saveSlot: number | undefined,
   gitRemoteURL: string | undefined,
   skipInstall: boolean,
   packageManager: PackageManager,
+  ts: boolean,
   dev: boolean,
   verbose: boolean,
 ): void {
@@ -48,15 +55,16 @@ export function createProject(
   }
 
   const config = new Config(modsDirectory, saveSlot, dev);
-  createConfigFile(projectPath, config, verbose);
+  createConfigFile(projectPath, config, ts, verbose);
 
-  copyStaticFiles(projectPath, verbose);
+  copyStaticFiles(projectPath, ts, verbose);
   copyDynamicFiles(
     projectName,
     authorName,
     projectPath,
     packageManager,
     dev,
+    ts,
     verbose,
   );
   updateNodeModules(projectPath, packageManager, verbose);
@@ -66,21 +74,25 @@ export function createProject(
   // Only make the initial commit once all of the files have been copied and formatted.
   initGitRepository(projectPath, gitRemoteURL, verbose);
 
-  console.log(`Successfully created mod: ${chalk.green(projectName)}`);
+  const noun = ts ? "project" : "mod";
+  console.log(`Successfully created ${noun}: ${chalk.green(projectName)}`);
 }
 
 /** Copy static files, like ".eslintrc.cjs", "tsconfig.json", etc. */
-function copyStaticFiles(projectPath: string, verbose: boolean) {
-  const staticFileList = file.getDirList(TEMPLATES_STATIC_DIR, verbose);
-  for (const fileName of staticFileList) {
-    const templateFilePath = path.join(TEMPLATES_STATIC_DIR, fileName);
-    const destinationFilePath = path.join(projectPath, fileName);
-    if (!file.exists(destinationFilePath, verbose)) {
-      file.copy(templateFilePath, destinationFilePath, verbose);
-    }
-  }
+function copyStaticFiles(projectPath: string, ts: boolean, verbose: boolean) {
+  // First, copy the static files that are shared between TypeScript projects and IsaacScript mods.
+  copyTemplateDirectoryWithoutOverwriting(
+    TEMPLATES_STATIC_DIR,
+    projectPath,
+    verbose,
+  );
 
-  // Rename ".eslintrc.template.js" to ".eslintrc.cjs". (If it is kept as ".eslintrc.cjs", then
+  // Second, copy files that are specific to either a TypeScript project or an IsaacScript mod.
+  const staticDirSuffix = ts ? "ts" : "mod";
+  const staticDirPath = path.join(TEMPLATES_DIR, `static-${staticDirSuffix}`);
+  copyTemplateDirectoryWithoutOverwriting(staticDirPath, projectPath, verbose);
+
+  // Rename ".eslintrc.template.cjs" to ".eslintrc.cjs". (If it is kept as ".eslintrc.cjs", then
   // local linting will fail.)
   const ESLintConfigPath = path.join(projectPath, ".eslintrc.template.cjs");
   const correctESLintConfigPath = path.join(projectPath, ".eslintrc.cjs");
@@ -99,6 +111,22 @@ function copyStaticFiles(projectPath: string, verbose: boolean) {
   file.rename(cSpellConfigPath, correctCSpellConfigPath, verbose);
 }
 
+function copyTemplateDirectoryWithoutOverwriting(
+  templateDirPath: string,
+  projectPath: string,
+  verbose: boolean,
+) {
+  const dirList = file.getDirList(templateDirPath, verbose);
+
+  for (const fileName of dirList) {
+    const templateFilePath = path.join(templateDirPath, fileName);
+    const destinationFilePath = path.join(projectPath, fileName);
+    if (!file.exists(destinationFilePath, verbose)) {
+      file.copy(templateFilePath, destinationFilePath, verbose);
+    }
+  }
+}
+
 /** Copy files that need to have text replaced inside of them. */
 function copyDynamicFiles(
   projectName: string,
@@ -106,6 +134,7 @@ function copyDynamicFiles(
   projectPath: string,
   packageManager: PackageManager,
   dev: boolean,
+  ts: boolean,
   verbose: boolean,
 ) {
   const workflowsPath = path.join(projectPath, ".github", "workflows");
@@ -117,13 +146,17 @@ function copyDynamicFiles(
     const templatePath = CI_YML_TEMPLATE_PATH;
     const template = file.read(templatePath, verbose);
 
+    // There are two versions of the template, one for TypeScript, and one for IsaacScript mods.
+    const modifiedTemplate = ts
+      ? removeLinesBetweenMarkers(template, TEMPLATE_MOD_MARKER)
+      : removeLinesMatching(template, TEMPLATE_MOD_MARKER);
+
     const lockFileName = getPackageManagerLockFileName(packageManager);
     const installCommand = getPackageManagerInstallCICommand(packageManager);
-    const ciYML = template
-      .replaceAll("PACKAGE-MANAGER-NAME", packageManager)
-      .replaceAll("PACKAGE-MANAGER-LOCK-FILE-NAME", lockFileName)
-      .replaceAll("PACKAGE-MANAGER-INSTALL", installCommand)
-      .replaceAll("PROJECT-NAME", projectName);
+    const ciYML = modifiedTemplate
+      .replaceAll("PACKAGE_MANAGER_NAME", packageManager)
+      .replaceAll("PACKAGE_MANAGER_LOCK_FILE_NAME", lockFileName)
+      .replaceAll("PACKAGE_MANAGER_INSTALL_COMMAND", installCommand);
 
     const destinationPath = path.join(workflowsPath, fileName);
     file.write(destinationPath, ciYML, verbose);
@@ -135,6 +168,11 @@ function copyDynamicFiles(
     const templatePath = GITIGNORE_TEMPLATE_PATH;
     const template = file.read(templatePath, verbose);
 
+    // There are two versions of the template, one for TypeScript, and one for IsaacScript mods.
+    const modifiedTemplate = ts
+      ? removeLinesBetweenMarkers(template, TEMPLATE_MOD_MARKER)
+      : removeLinesMatching(template, TEMPLATE_MOD_MARKER);
+
     // Prepend a header with the project name.
     let separatorLine = "# ";
     repeat(projectName.length, () => {
@@ -142,46 +180,55 @@ function copyDynamicFiles(
     });
     separatorLine += "\n";
     const gitignoreHeader = `${separatorLine}# ${projectName}\n${separatorLine}\n`;
-    const gitignore = gitignoreHeader + template;
+    const gitignore = gitignoreHeader + modifiedTemplate;
 
-    const destinationPath = path.join(projectPath, `.${fileName}`); // We need to prepend a period
+    const destinationPath = path.join(projectPath, `.${fileName}`); // We need to prepend a period.
     file.write(destinationPath, gitignore, verbose);
   }
 
   // `package.json`
   {
-    // Modify and copy the file.
-    const fileName = PACKAGE_JSON;
-    const templatePath = PACKAGE_JSON_TEMPLATE_PATH;
+    // There are two versions of the template, one for TypeScript, and one for IsaacScript mods.
+    const packageJSONTemplateFileName = ts
+      ? "package.ts.json"
+      : "package.mod.json";
+    const templatePath = path.join(
+      TEMPLATES_DYNAMIC_DIR,
+      packageJSONTemplateFileName,
+    );
     const template = file.read(templatePath, verbose);
 
-    let packageJSON = template.replaceAll("MOD-NAME-TO-REPLACE", projectName);
-    if (authorName === undefined) {
-      // Remove the "author" line.
-      packageJSON = packageJSON.replace(/^\s*"author":.*$/gm, "");
-    } else {
-      packageJSON = packageJSON.replaceAll("AUTHOR-NAME", authorName);
-    }
+    const packageJSON = template
+      .replaceAll("PROJECT_NAME", projectName)
+      .replaceAll("AUTHOR_NAME", authorName ?? "unknown")
+      .replaceAll("PACKAGE_MANAGER", packageManager);
 
-    const destinationPath = path.join(projectPath, fileName);
+    const destinationPath = path.join(projectPath, PACKAGE_JSON);
     file.write(destinationPath, packageJSON, verbose);
   }
 
   // `README.md`
   {
-    const fileName = README_MD;
-    const templatePath = README_MD_TEMPLATES_PATH;
+    // There are two versions of the template, one for TypeScript, and one for IsaacScript mods.
+    const readmeMDTemplateFileName = ts ? "README.ts.json" : "README.mod.json";
+    const templatePath = path.join(
+      TEMPLATES_DYNAMIC_DIR,
+      readmeMDTemplateFileName,
+    );
     const template = file.read(templatePath, verbose);
-    const readmeMD = template.replaceAll("MOD-NAME-TO-REPLACE", projectName);
-    const destinationPath = path.join(projectPath, fileName);
+
+    // "PROJECT-NAME" must be hyphenated, as using an underscore will break Prettier for some
+    // reason.
+    const readmeMD = template.replaceAll("PROJECT-NAME", projectName);
+    const destinationPath = path.join(projectPath, README_MD);
     file.write(destinationPath, readmeMD, verbose);
   }
 
-  const modPath = path.join(projectPath, "mod");
-  file.makeDir(modPath, verbose);
-
   // `mod/metadata.xml`
-  {
+  if (!ts) {
+    const modPath = path.join(projectPath, "mod");
+    file.makeDir(modPath, verbose);
+
     const fileName = METADATA_XML;
     const templatePath = METADATA_XML_TEMPLATE_PATH;
     const template = file.read(templatePath, verbose);
@@ -193,8 +240,8 @@ function copyDynamicFiles(
   const srcPath = path.join(projectPath, "src");
   file.makeDir(srcPath, verbose);
 
-  // `src/main.ts`
-  {
+  // `src/main.ts` (TypeScript projects use the simple version from the "static-ts" directory.)
+  if (!ts) {
     // Convert snake_case and kebab-case to camelCase. (Kebab-case in particular will make the
     // example TypeScript file fail to compile.)
     const fileName = MAIN_TS;
@@ -207,7 +254,7 @@ function copyDynamicFiles(
 
   // If we are initializing an IsaacScript project intended to be used for development, we can
   // include a better starter file.
-  if (dev) {
+  if (!ts && dev) {
     const fileName = MAIN_TS;
     const templatePath = MAIN_DEV_TS_TEMPLATE_PATH;
     const template = file.read(templatePath, verbose);